napagana ko na rin yung edit/delete ng diary, ukininam

+++ app/src/main/java/com/bigo143/echodiary/DiaryDao.java
@@ -21,6 +21,6 @@ public interface DiaryDao {
     @Update
     void update(DiaryEntry entry);
     @Query("DELETE FROM diary_entries WHERE id = :id")
-    void deleteById(int id);
+    void deleteById(long id);
 
 }
